requirements.txt
requests==2.28.2
chardet==3.0.4
psutil==5.8.0
pyyaml==5.4
pytz==2021.1
ansible==2.9.6
pytest-testinfra==8.1.0